Move time.h and time_libultra.c to src/system folder.
