+++ src/effects/splash_particle_effect.c
@@ -1,7 +1,7 @@
 #include "splash_particle_effect.h"
 
 #include "../math/mathf.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../math/vector2.h"
 #include "../physics/config.h"
 #include "../scene/dynamic_scene.h"
