+++ src/levels/credits.c
@@ -1,6 +1,6 @@
 #include "credits.h"
 
-#include "../util/time.h"
+#include "system/time.h"
 #include "./levels.h"
 #include "../audio/soundplayer.h"
 #include "../build/src/audio/clips.h"
