+++ src/physics/point_constraint.c
@@ -1,6 +1,6 @@
 #include "point_constraint.h"
 
-#include "../util/time.h"
+#include "system/time.h"
 
 #define BREAK_CONSTRAINT_DISTANCE 1.25f
 #define CLAMP_CONSTRAINT_DISTANCE 0.07f
