+++ src/physics/rigid_body.c
@@ -1,5 +1,5 @@
 #include "rigid_body.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../physics/config.h"
 #include "contact_solver.h"
 #include "collision_scene.h"
