+++ src/savefile/savefile.c
@@ -1,7 +1,7 @@
 
 #include "savefile.h"
 #include "util/memory.h"
-#include "util/time.h"
+#include "system/time.h"
 #include "system/controller.h"
 
 #include "../controls/controller_actions.h"
