+++ src/scene/ball.c
@@ -6,7 +6,7 @@
 #include "../physics/collision_scene.h"
 #include "../physics/collision_box.h"
 
-#include "../util/time.h"
+#include "system/time.h"
 
 #include "../build/assets/models/grav_flare.h"
 #include "../build/assets/models/fleck_ash2.h"
