+++ src/scene/box_dropper.c
@@ -3,7 +3,7 @@
 #include "../scene/dynamic_scene.h"
 #include "../defs.h"
 #include "../physics/config.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "signals.h"
 #include "../decor/decor_object_list.h"
 #include "../util/dynamic_asset_loader.h"
