+++ src/scene/button.c
@@ -6,7 +6,7 @@
 #include "../physics/collision_cylinder.h"
 #include "../physics/collision_scene.h"
 #include "../physics/contact_solver.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../util/dynamic_asset_loader.h"
 
 #include "../../build/assets/models/dynamic_animated_model_list.h"
