+++ src/scene/door.c
@@ -5,7 +5,7 @@
 #include "../scene/dynamic_scene.h"
 #include "signals.h"
 #include "../math/mathf.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../physics/collision_box.h"
 #include "../physics/collision_scene.h"
 #include "../util/dynamic_asset_loader.h"
