+++ src/scene/elevator.c
@@ -4,7 +4,7 @@
 #include "../physics/mesh_collider.h"
 #include "./signals.h"
 #include "../math/mathf.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../audio/soundplayer.h"
 #include "../audio/clips.h"
 #include "../controls/rumble_pak.h"
