+++ src/scene/hud.c
@@ -3,7 +3,7 @@
 #include "../../build/assets/materials/hud.h"
 #include "../menu/controls.h"
 #include "../graphics/graphics.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../levels/levels.h"
 #include "./scene.h"
 #include "../savefile/savefile.h"
