+++ src/scene/portal.c
@@ -7,7 +7,7 @@
 #include "../physics/collision_scene.h"
 #include "../math/mathf.h"
 #include "../math/vector2s16.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../levels/levels.h"
 #include "./portal_surface_generator.h"
 #include "../scene/dynamic_scene.h"
