+++ src/scene/scene.c
@@ -7,7 +7,7 @@
 #include "materials/subject.h"
 #include "materials/light.h"
 #include "materials/point_light_rendered.h"
-#include "util/time.h"
+#include "system/time.h"
 #include "sk64/skelatool_defs.h"
 #include "shadow_map.h"
 #include "../physics/point_constraint.h"
