+++ src/scene/scene_animator.c
@@ -1,7 +1,7 @@
 #include "scene_animator.h"
 
 #include "../util/memory.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../math/mathf.h"
 #include "../defs.h"
