+++ src/scene/security_camera.c
@@ -5,7 +5,7 @@
 #include "../physics/collision_scene.h"
 #include "dynamic_scene.h"
 #include "scene.h"
-#include "../util/time.h"
+#include "system/time.h"
 #include "../levels/cutscene_runner.h"
 #include "../decor/decor_object.h"
 #include "../util/dynamic_asset_loader.h"
