+++ src/scene/signage.c
@@ -4,7 +4,7 @@
 #include "../levels/levels.h"
 #include "../defs.h"
 #include "../graphics/color.h"
-#include "../util/time.h"
+#include "system/time.h"
 
 #include "../build/assets/models/props/signage.h"
 #include "../build/assets/models/props/cylinder_test.h"
