+++ src/scene/switch.c
@@ -10,7 +10,7 @@
 #include "../build/assets/materials/static.h"
 #include "../../build/assets/models/dynamic_animated_model_list.h"
 
-#include "../util/time.h"
+#include "system/time.h"
 #include "../scene/hud.h"
 #include "../scene/scene.h"
